refactor: simplify version control module folder structure
