+++ src/modules/version-control/automerge-repo/browser/index.ts
@@ -1,0 +1,1 @@
+export { setupForWeb, setupForElectron } from './setup';
