+++ src/modules/version-control/automerge-repo/electron-ipc-network-adapter/index.ts
