+++ src/modules/version-control/index.ts
@@ -1,4 +1,4 @@
-export * from './repo/common';
+export * from './automerge-repo/common';
 export * from './models';
 export type { VersionControlRepo } from './ports/version-control-repo';
 export * from './commands';
