--- src/modules/version-control/repo/browser/index.ts
@@ -1,6 +1,0 @@
-export { setupForWeb, setupForElectron } from './setup';
-
-export { useHandle } from './react/useHandle';
-export { useDocument } from './react/useDocument';
-
-export { VersionControlContext, VersionControlProvider } from './react/context';
