--- src/modules/version-control/repo/browser/react/index.ts
@@ -1,2 +1,0 @@
-export { useDocument } from './useDocument';
-export { useHandle } from './useHandle';
