--- src/modules/version-control/repo/browser/react/useDocument.ts
@@ -1,1 +1,0 @@
-export { useDocument } from '@automerge/automerge-repo-react-hooks';
