--- src/modules/version-control/repo/browser/react/useHandle.ts
@@ -1,21 +1,0 @@
-import { AutomergeUrl } from '@automerge/automerge-repo/slim';
-import { useHandle as useHandleAutomerge } from '@automerge/automerge-repo-react-hooks';
-import { useEffect, useState } from 'react';
-
-import { RichTextDocument } from '../../../models/document';
-
-export function useHandle(docUrl: AutomergeUrl) {
-  const handle = useHandleAutomerge<RichTextDocument>(docUrl);
-  const [isReady, setIsReady] = useState(false);
-
-  useEffect(() => {
-    if (handle && !isReady) {
-      handle.whenReady().then(() => {
-        setIsReady(true);
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [handle]);
-
-  return { handle, isReady };
-}
